LayoutTests/platform/qt-4.8/fast/xsl/subframe-location-expected.txt

CONSOLE MESSAGE: line 4: <html xmlns='http://www.w3.org/1999/xhtml/'><body><p>Running an XSL-T 1.0 stylesheet with a 2.0 processor.</p></body></html>
ALERT: SUCCESS
Test that a document in a subframe processed with XSLT has a correct location. (bug 18259).
